decoder: append data to fifo when rx valid received
Used my_mipi_rx_valid to get if the mipi frame is in active mode
or not, if there is active frame then and only then data will be
stored to fifo.

Also, Data len of the frame will checked while aligning and if
data len doesn't match then from there the fifo will be in reset
mode.